+++ terra_geocrud/tests/test_properties/test_widgets.py
@@ -19,37 +19,6 @@ class BaseWidgetTestCase(TestCase):
             widget.render()
 
 
-class BaseDataFileWidgetTestCase(TestCase):
-    def setUp(self) -> None:
-        self.property_key = 'logo'
-        self.feature_with_file_name = FeatureFactory(
-            properties={
-                self.property_key: 'data:image/png;name=toto.png;base64,xxxxxxxxxxxxxxxxxxxxxxxxxx=='
-            }
-        )
-        self.feature_without_file_name = FeatureFactory(
-            properties={
-                self.property_key: 'data:image/png;base64,xxxxxxxxxxxxxxxxxxxxxxxxxx=='
-            }
-        )
-        self.feature_without_file_data = FeatureFactory(
-            properties={
-                self.property_key: None
-            }
-        )
-
-    def test_get_info_content_no_data(self):
-        widget = widgets.BaseDataFileWidget(feature=self.feature_without_file_data, prop=self.property_key)
-        info, content = widget.get_info_content()
-        self.assertIsNone(info)
-        self.assertIsNone(content)
-
-    def test_get_info_content_no_file_name(self):
-        widget = widgets.BaseDataFileWidget(feature=self.feature_without_file_name, prop=self.property_key)
-        path = widget.get_storage_file_path()
-        self.assertTrue(path.endswith(f'{self.property_key}.png'), path)
-
-
 class GetWidgetChoicesTestCase(TestCase):
     def setUp(self) -> None:
         self.choices = widgets.get_widgets_choices()